chore: rename interfaces
