+++ src/lib/nodemailer.ts
@@ -30,6 +30,6 @@ export async function sendEmail(
 
     await transporter.sendMail(mailOptions);
   } catch (error: any) {
-    console.error(`Erro ao enviar e-mail: ${error}`);
+    console.error(`Error to send e-mail: ${error}`);
   }
 }
