+++ src/shared/interfaces/Houses.ts
