+++ src/shared/interfaces/PlayerData.ts
